+++ examples/ota-provider-app/linux/main.cpp
@@ -28,7 +28,6 @@
 #include <lib/core/CHIPError.h>
 #include <lib/support/CHIPArgParser.hpp>
 #include <lib/support/CHIPMem.h>
-#include <lib/support/RandUtils.h>
 #include <lib/support/logging/CHIPLogging.h>
 
 #include <ota-provider-common/BdxOtaSender.h>
