+++ examples/shell/standalone/main.cpp
@@ -21,7 +21,6 @@
 #include <lib/support/Base64.h>
 #include <lib/support/CHIPArgParser.hpp>
 #include <lib/support/CodeUtils.h>
-#include <lib/support/RandUtils.h>
 #include <lib/support/logging/CHIPLogging.h>
 
 #include <ChipShellCollection.h>
