+++ src/crypto/RandUtils.cpp
@@ -18,7 +18,7 @@
 
 /**
  *    @file
- *      This file defines utility functions for deriving random integers.
+ *      This file implements utility functions for deriving random integers.
  *
  *  @note These utility functions do not generate cryptographically strong
  *        random number. To get cryptographically strong random data use
@@ -26,42 +26,44 @@
  *
  */
 
-#pragma once
+#include "RandUtils.h"
 
 #include <stdint.h>
+#include <stdlib.h>
+
+#include <crypto/CHIPCryptoPAL.h>
+#include <lib/support/CodeUtils.h>
 
 namespace chip {
+namespace Crypto {
 
-/**
- *  This function generates 64-bit unsigned random number.
- *
- * @return  64-bit unsigned random integer.
- *
- */
-extern uint64_t GetRandU64();
+uint64_t GetRandU64()
+{
+    uint64_t tmp = 0;
+    VerifyOrDie(CHIP_NO_ERROR == DRBG_get_bytes(reinterpret_cast<uint8_t *>(&tmp), sizeof(tmp)));
+    return tmp;
+}
 
-/**
- *  This function generates 32-bit unsigned random number.
- *
- * @return  32-bit unsigned random integer.
- *
- */
-extern uint32_t GetRandU32();
+uint32_t GetRandU32()
+{
+    uint32_t tmp = 0;
+    VerifyOrDie(CHIP_NO_ERROR == DRBG_get_bytes(reinterpret_cast<uint8_t *>(&tmp), sizeof(tmp)));
+    return tmp;
+}
 
-/**
- *  This function generates 16-bit unsigned random number.
- *
- * @return  16-bit unsigned random integer.
- *
- */
-extern uint16_t GetRandU16();
+uint16_t GetRandU16()
+{
+    uint16_t tmp = 0;
+    VerifyOrDie(CHIP_NO_ERROR == DRBG_get_bytes(reinterpret_cast<uint8_t *>(&tmp), sizeof(tmp)));
+    return tmp;
+}
 
-/**
- *  This function generates 8-bit unsigned random number.
- *
- * @return  8-bit unsigned random integer.
- *
- */
-extern uint8_t GetRandU8();
+uint8_t GetRandU8()
+{
+    uint8_t tmp = 0;
+    VerifyOrDie(CHIP_NO_ERROR == DRBG_get_bytes(&tmp, sizeof(tmp)));
+    return tmp;
+}
 
+} // namespace Crypto
 } // namespace chip
