+++ src/lib/dnssd/BUILD.gn
@@ -23,6 +23,7 @@ source_set("platform_header") {
 static_library("dnssd") {
   public_deps = [
     ":platform_header",
+    "${chip_root}/src/crypto",
     "${chip_root}/src/lib/core",
     "${chip_root}/src/lib/support",
   ]
