+++ src/lib/support/BUILD.gn
@@ -88,8 +88,6 @@ static_library("support") {
     "Pool.h",
     "PrivateHeap.cpp",
     "PrivateHeap.h",
-    "RandUtils.cpp",
-    "RandUtils.h",
     "ReferenceCountedHandle.h",
     "SafeInt.h",
     "SerializableIntegerSet.cpp",
