+++ src/transport/MessageCounter.cpp
@@ -22,12 +22,12 @@
 
 #include <transport/MessageCounter.h>
 
-#include <lib/support/RandUtils.h>
+#include <crypto/RandUtils.h>
 #include <platform/CHIPDeviceLayer.h>
 
 namespace chip {
 
-GlobalUnencryptedMessageCounter::GlobalUnencryptedMessageCounter() : value(GetRandU32()) {}
+GlobalUnencryptedMessageCounter::GlobalUnencryptedMessageCounter() : value(Crypto::GetRandU32()) {}
 
 CHIP_ERROR GlobalEncryptedMessageCounter::Init()
 {
